fixed names to follow UpperCamelCase convention, renamed QCode to PolishQCode, some fixes in directories
