+++ src/Entity/RSTReports/Polish/Readability.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Entity\RST_Reports\Polish;
+namespace App\Entity\RSTReports\Polish;
 
-use App\Entity\RST_Reports\Interfaces\ReadablilityInterface;
+use App\Entity\RSTReports\Interfaces\ReadablilityInterface;
 
 enum Readability: string implements ReadablilityInterface
 {
