--- src/Entity/RST_Reports/Interfaces/ToneInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Entity\RST_Reports\Interfaces;
-
-interface ToneInterface
-{
-
-}
