--- src/Entity/RST_Reports/RST_Report.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace App\Entity\RST_Reports;
-
-use App\Entity\RST_Reports\Interfaces\ModulationInterface;
-use App\Entity\RST_Reports\Interfaces\ReadablilityInterface;
-use App\Entity\RST_Reports\Interfaces\StrengthInterface;
-use App\Entity\RST_Reports\Interfaces\ToneInterface;
-
-class RST_Report
-{
-    public ReadablilityInterface $readability;
-    public StrengthInterface $strength;
-    public ToneInterface $tone;
-    public ModulationInterface $modulation;
-
-}
